final version of fractol, with color modes and presets, multi-threeading, iteration coefficient through zoom.

+++ libft/ft_strncpy.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_strncpy.c                                       :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: aditsch <aditsch@student.42.fr>            +#+  +:+       +#+        */
+/*   By: gpoblon <gpoblon@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2016/09/11 01:28:39 by aditsch           #+#    #+#             */
-/*   Updated: 2016/11/09 11:01:36 by aditsch          ###   ########.fr       */
+/*   Created: 2016/09/23 16:46:21 by gpoblon           #+#    #+#             */
+/*   Updated: 2016/11/04 12:58:32 by gpoblon          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 char	*ft_strncpy(char *dst, const char *src, size_t n)
 {
-	unsigned int	i;
+	size_t i;
 
 	i = 0;
 	while (i < n && src[i])
